Defensive programming to prevent integer overflow for large images
Sometimes for very large images, either the number of pixels (w * h * d)
or the total data size (npixels * pixelbytes) can overflow int32 math.
Find several places where this may be a problem and use casts to ensure
64 bit multiplication.

I don't claim this is a comprehensive fix, but I did a cursory check of
some obvious places and found these to squash.

+++ src/dpx.imageio/libdpx/RunLengthEncoding.cpp
@@ -123,7 +123,7 @@ bool dpx::RunLengthEncoding::Read(const Header &dpxHeader, ElementReadStream *fd
 
 
 		// size of the image
-		const size_t imageSize = width * height * numberOfComponents;
+		const size_t imageSize = size_t(width) * size_t(height) * numberOfComponents;
 		const size_t imageByteSize = imageSize * byteCount;
 		
 		// allocate the buffer that will store the entire image
